split out Unison.Syntax.DeclParser from Unison.Syntax.FileParser
