+++ parser-typechecker/src/Unison/Syntax/FileParser.hs
@@ -7,7 +7,7 @@ import Data.Map qualified as Map
 import Data.Set qualified as Set
 import Text.Megaparsec qualified as P
 import Unison.ABT qualified as ABT
-import Unison.DataDeclaration (DataDeclaration, EffectDeclaration)
+import Unison.DataDeclaration (DataDeclaration)
 import Unison.DataDeclaration qualified as DD
 import Unison.Name qualified as Name
 import Unison.Names qualified as Names
@@ -15,15 +15,13 @@ import Unison.Names.ResolutionResult qualified as Names
 import Unison.NamesWithHistory qualified as NamesWithHistory
 import Unison.Parser.Ann (Ann)
 import Unison.Prelude
+import Unison.Syntax.DeclParser (declarations)
 import Unison.Syntax.Lexer qualified as L
-import Unison.Syntax.Name qualified as Name (toText, toVar, unsafeFromVar)
+import Unison.Syntax.Name qualified as Name (toVar, unsafeFromVar)
 import Unison.Syntax.Parser
 import Unison.Syntax.TermParser qualified as TermParser
-import Unison.Syntax.TypeParser qualified as TypeParser
 import Unison.Term (Term)
 import Unison.Term qualified as Term
-import Unison.Type (Type)
-import Unison.Type qualified as Type
 import Unison.UnisonFile (UnisonFile (..))
 import Unison.UnisonFile qualified as UF
 import Unison.UnisonFile.Env qualified as UF
@@ -239,196 +237,3 @@ watched = P.try do
   tok <- anyToken
   guard $ maybe True (`L.touches` tok) kind
   pure (maybe UF.RegularWatch L.payload kind, guid, maybe mempty ann kind <> ann tok)
-
--- The parsed form of record accessors, as in:
---
--- type Additive a = { zero : a, (+) : a -> a -> a }
---
--- The `Token v` is the variable name and location (here `zero` and `(+)`) of
--- each field, and the type is the type of that field
-type Accessors v = [(L.Token v, [(L.Token v, Type v Ann)])]
-
-declarations ::
-  (Monad m, Var v) =>
-  P
-    v
-    m
-    ( Map v (DataDeclaration v Ann),
-      Map v (EffectDeclaration v Ann),
-      Accessors v
-    )
-declarations = do
-  declarations <- many $ declaration <* optional semi
-  let (dataDecls0, effectDecls) = partitionEithers declarations
-      dataDecls = [(a, b) | (a, b, _) <- dataDecls0]
-      multimap :: (Ord k) => [(k, v)] -> Map k [v]
-      multimap = foldl' mi Map.empty
-      mi m (k, v) = Map.insertWith (++) k [v] m
-      mds = multimap dataDecls
-      mes = multimap effectDecls
-      mdsBad = Map.filter (\xs -> length xs /= 1) mds
-      mesBad = Map.filter (\xs -> length xs /= 1) mes
-  if Map.null mdsBad && Map.null mesBad
-    then
-      pure
-        ( Map.fromList dataDecls,
-          Map.fromList effectDecls,
-          join . map (view _3) $ dataDecls0
-        )
-    else
-      P.customFailure . DuplicateTypeNames $
-        [(v, DD.annotation <$> ds) | (v, ds) <- Map.toList mdsBad]
-          <> [(v, DD.annotation . DD.toDataDecl <$> es) | (v, es) <- Map.toList mesBad]
-
--- unique[someguid] type Blah = ...
-modifier :: (Monad m, Var v) => P v m (Maybe (L.Token DD.Modifier))
-modifier = do
-  optional (unique <|> structural)
-  where
-    unique = do
-      tok <- openBlockWith "unique"
-      uid <- do
-        o <- optional (openBlockWith "[" *> wordyIdString <* closeBlock)
-        case o of
-          Nothing -> uniqueName 32
-          Just uid -> pure (fromString . L.payload $ uid)
-      pure (DD.Unique uid <$ tok)
-    structural = do
-      tok <- openBlockWith "structural"
-      pure (DD.Structural <$ tok)
-
-declaration ::
-  (Monad m, Var v) =>
-  P
-    v
-    m
-    ( Either
-        (v, DataDeclaration v Ann, Accessors v)
-        (v, EffectDeclaration v Ann)
-    )
-declaration = do
-  mod <- modifier
-  fmap Right (effectDeclaration mod) <|> fmap Left (dataDeclaration mod)
-
-dataDeclaration ::
-  forall m v.
-  (Monad m, Var v) =>
-  Maybe (L.Token DD.Modifier) ->
-  P v m (v, DataDeclaration v Ann, Accessors v)
-dataDeclaration mod = do
-  keywordTok <- fmap void (reserved "type") <|> openBlockWith "type"
-  (name, typeArgs) <-
-    (,)
-      <$> TermParser.verifyRelativeVarName prefixDefinitionName
-      <*> many (TermParser.verifyRelativeVarName prefixDefinitionName)
-  let typeArgVs = L.payload <$> typeArgs
-  eq <- reserved "="
-  let -- go gives the type of the constructor, given the types of
-      -- the constructor arguments, e.g. Cons becomes forall a . a -> List a -> List a
-      go :: L.Token v -> [Type v Ann] -> (Ann, v, Type v Ann)
-      go ctorName ctorArgs =
-        let arrow i o = Type.arrow (ann i <> ann o) i o
-            app f arg = Type.app (ann f <> ann arg) f arg
-            -- ctorReturnType e.g. `Optional a`
-            ctorReturnType = foldl' app (tok Type.var name) (tok Type.var <$> typeArgs)
-            -- ctorType e.g. `a -> Optional a`
-            --    or just `Optional a` in the case of `None`
-            ctorType = foldr arrow ctorReturnType ctorArgs
-            ctorAnn = ann ctorName <> maybe (ann ctorName) ann (lastMay ctorArgs)
-         in ( ann ctorName,
-              Var.namespaced [L.payload name, L.payload ctorName],
-              Type.foralls ctorAnn typeArgVs ctorType
-            )
-      prefixVar = TermParser.verifyRelativeVarName prefixDefinitionName
-      dataConstructor :: P v m (Ann, v, Type v Ann)
-      dataConstructor = go <$> prefixVar <*> many TypeParser.valueTypeLeaf
-      record :: P v m ([(Ann, v, Type v Ann)], [(L.Token v, [(L.Token v, Type v Ann)])])
-      record = do
-        _ <- openBlockWith "{"
-        let field :: P v m [(L.Token v, Type v Ann)]
-            field = do
-              f <- liftA2 (,) (prefixVar <* reserved ":") TypeParser.valueType
-              optional (reserved ",")
-                >>= ( \case
-                        Nothing -> pure [f]
-                        Just _ -> maybe [f] (f :) <$> (optional semi *> optional field)
-                    )
-        fields <- field
-        _ <- closeBlock
-        let lastSegment = name <&> (\v -> Var.named (Name.toText $ Name.unqualified (Name.unsafeFromVar v)))
-        pure ([go lastSegment (snd <$> fields)], [(name, fields)])
-  (constructors, accessors) <-
-    msum [record, (,[]) <$> sepBy (reserved "|") dataConstructor]
-  _ <- closeBlock
-  let -- the annotation of the last constructor if present,
-      -- otherwise ann of name
-      closingAnn :: Ann
-      closingAnn = last (ann eq : ((\(_, _, t) -> ann t) <$> constructors))
-  case mod of
-    Nothing -> P.customFailure $ MissingTypeModifier ("type" <$ keywordTok) name
-    Just mod' ->
-      pure
-        ( L.payload name,
-          DD.mkDataDecl' (L.payload mod') (ann mod' <> closingAnn) typeArgVs constructors,
-          accessors
-        )
-
-effectDeclaration ::
-  forall m v.
-  (Monad m, Var v) =>
-  Maybe (L.Token DD.Modifier) ->
-  P v m (v, EffectDeclaration v Ann)
-effectDeclaration mod = do
-  keywordTok <- fmap void (reserved "ability") <|> openBlockWith "ability"
-  name <- TermParser.verifyRelativeVarName prefixDefinitionName
-  typeArgs <- many (TermParser.verifyRelativeVarName prefixDefinitionName)
-  let typeArgVs = L.payload <$> typeArgs
-  blockStart <- openBlockWith "where"
-  constructors <- sepBy semi (constructor typeArgs name)
-  -- `ability` opens a block, as does `where`
-  _ <- closeBlock <* closeBlock
-  let closingAnn =
-        last $ ann blockStart : ((\(_, _, t) -> ann t) <$> constructors)
-
-  case mod of
-    Nothing -> P.customFailure $ MissingTypeModifier ("ability" <$ keywordTok) name
-    Just mod' ->
-      pure
-        ( L.payload name,
-          DD.mkEffectDecl'
-            (L.payload mod')
-            (ann mod' <> closingAnn)
-            typeArgVs
-            constructors
-        )
-  where
-    constructor :: [L.Token v] -> L.Token v -> P v m (Ann, v, Type v Ann)
-    constructor typeArgs name =
-      explodeToken
-        <$> TermParser.verifyRelativeVarName prefixDefinitionName
-        <* reserved ":"
-        <*> ( Type.generalizeLowercase mempty
-                . ensureEffect
-                <$> TypeParser.computationType
-            )
-      where
-        explodeToken v t = (ann v, Var.namespaced [L.payload name, L.payload v], t)
-        -- If the effect is not syntactically present in the constructor types,
-        -- add them after parsing.
-        ensureEffect t = case t of
-          Type.Effect' _ _ -> modEffect t
-          x -> Type.editFunctionResult modEffect x
-        modEffect t = case t of
-          Type.Effect' es t -> go es t
-          t -> go [] t
-        toTypeVar t = Type.av' (ann t) (Var.name $ L.payload t)
-        headIs t v = case t of
-          Type.Apps' (Type.Var' x) _ -> x == v
-          Type.Var' x -> x == v
-          _ -> False
-        go es t =
-          let es' =
-                if any (`headIs` L.payload name) es
-                  then es
-                  else Type.apps' (toTypeVar name) (toTypeVar <$> typeArgs) : es
-           in Type.cleanupAbilityLists $ Type.effect (ABT.annotation t) es' t
